requirements.txt
asgiref==3.7.2
boto3==1.28.82
botocore==1.31.82
crispy-bootstrap4==2022.1
distlib==0.3.6
Django==4.2.3
django-crispy-forms==2.0
django-filter==23.2
djangorestframework==3.14.0
django-storages==1.14.2
filelock==3.12.2
jmespath==1.0.1
Pillow==10.0.0
platformdirs==3.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.7
virtualenv==20.23.1
virtualenvwrapper-win==1.2.7
